# backend/docs/swagger.yaml
basePath: /
definitions:
  models.StudentScore:
    properties:
      diaLi:
        type: number
      gdcd:
        type: number
      hoaHoc:
        type: number
      id:
        type: integer
      lichSu:
        type: number
      maNgoaiNgu:
        type: string
      ngoaiNgu:
        type: number
      nguVan:
        type: number
      sbd:
        type: string
      sinhHoc:
        type: number
      toan:
        type: number
      vatLi:
        type: number
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is an API for student score management.
  title: GoldenOwl Student Score API
  version: "1.0"
paths:
  /student-scores:
    get:
      description: Returns student score based on registration number (SBD)
      parameters:
      - description: Student Registration Number
        in: query
        name: sbd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StudentScore'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get student score by SBD
      tags:
      - student-scores
  /student-scores/report:
    get:
      description: Returns an aggregated report of scores by subject
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StudentScore'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get score report by subject
      tags:
      - student-scores
  /student-scores/top-10-groupA:
    get:
      description: Returns the top 10 students in Group A based on their scores
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StudentScore'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get top 10 scores for Group A
      tags:
      - student-scores
swagger: "2.0"